move contract source identifier to compiler module
