+++ protostar/compiler/__init__.py
@@ -11,3 +11,7 @@ from .contract_identifier_resolver import (
     ContractIdentifierResolver,
     ContractIdentificationException,
 )
+from .contract_source_identifier import (
+    ContractSourceIdentifier,
+    create_contract_source_identifier_factory,
+)
